task_unpacker/requirements.txt
streamlit
langchain
openai
python-dotenv
pymupdf
fpdf
python-docx